% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ncbi_superfamily_Ani.R
\name{get_ncbi_superfamily_Ani}
\alias{get_ncbi_superfamily_Ani}
\title{批量获取动物科级（Family）对应的 NCBI 总科（Superfamily）信息}
\usage{
get_ncbi_superfamily_Ani(family_vector)
}
\arguments{
\item{family_vector}{字符型向量 (Character vector)。包含需要查询的“科”名组合（例如 \code{c("Noctuidae", "Geometridae")}）。}
}
\value{
返回一个数据框 (Data.frame)。必定包含两列：
\code{family}（输入的科名）和 \code{NCBI.superfamily}（在 NCBI 中查到的总科名）。
如果某个科名在 NCBI 中未查到，对应的总科列将填充为 \code{NA}。
}
\description{
#' @description
专为动物/昆虫数据集的“总科补齐”模块。
由于 GBIF 和 BOLD 数据库的分类树通常缺失“总科 (superfamily)”这一分类阶元，
该函数通过提取数据中已查到的“科 (family)”名，前往 NCBI 数据库进行批量反向查询，
拉取完整的分类树，并剥离出对应的 NCBI 总科名称。
}
